+++ scripts/test/DynamicField/SearchSQLGet.t
@@ -155,6 +155,15 @@ my %DynamicFieldConfigs = (
     },
 );
 
+# Set expected results depends on database case sensitivity (bug#12657).
+my $SearchTerm = "(\'Foo\')";
+my $ValueText  = '(dfv.value_text)';
+
+if ( $DBObject->GetDatabaseFunction('CaseSensitive') ) {
+    $ValueText  = 'LOWER' . $ValueText;
+    $SearchTerm = 'LOWER' . $SearchTerm;
+}
+
 # define tests
 my @Tests = (
     {
@@ -219,14 +228,14 @@ my @Tests = (
             },
         },
         ExpectedResult => {
-            Equals            => " dfv.value_text = 'Foo' ",
-            GreaterThan       => " dfv.value_text > 'Foo' ",
-            GreaterThanEquals => " dfv.value_text >= 'Foo' ",
+            Equals            => " $ValueText = $SearchTerm ",
+            GreaterThan       => " $ValueText > $SearchTerm ",
+            GreaterThanEquals => " $ValueText >= $SearchTerm ",
             Like              => {
                 ColumnKey => 'dfv.value_text',
             },
-            SmallerThan       => " dfv.value_text < 'Foo' ",
-            SmallerThanEquals => " dfv.value_text <= 'Foo' ",
+            SmallerThan       => " $ValueText < $SearchTerm ",
+            SmallerThanEquals => " $ValueText <= $SearchTerm ",
         },
     },
     {
@@ -244,14 +253,14 @@ my @Tests = (
             },
         },
         ExpectedResult => {
-            Equals            => " dfv.value_text = 'Foo' ",
-            GreaterThan       => " dfv.value_text > 'Foo' ",
-            GreaterThanEquals => " dfv.value_text >= 'Foo' ",
+            Equals            => " $ValueText = $SearchTerm ",
+            GreaterThan       => " $ValueText > $SearchTerm ",
+            GreaterThanEquals => " $ValueText >= $SearchTerm ",
             Like              => {
                 ColumnKey => 'dfv.value_text',
             },
-            SmallerThan       => " dfv.value_text < 'Foo' ",
-            SmallerThanEquals => " dfv.value_text <= 'Foo' ",
+            SmallerThan       => " $ValueText < $SearchTerm ",
+            SmallerThanEquals => " $ValueText <= $SearchTerm ",
         },
     },
     {